third_party/blink/web_tests/http/tests/local/formdata/send-form-data-with-bad-string-expected.txt

Test that passing bad strings to FormData.append() throws and aborts properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS formData.append(badString, 'test') threw exception Exception in toString().
PASS formData.append('textarea', badString) threw exception Exception in toString().
PASS formData.append('blob', new Blob(['']), badString) threw exception Exception in toString().
PASS xhr.response.indexOf('test') is -1
PASS xhr.response.indexOf('textarea') is -1
PASS xhr.response.indexOf('blob') is -1
PASS successfullyParsed is true

TEST COMPLETE
